+++ src/Capital_Good.cpp
@@ -3,7 +3,7 @@
 
 /* Constructor
 */
-Capital_Good::Capital_Good(Capital_Firm_Agent *  seller_pointer, float seller_price, int seller_quantity,int life_span)
+Capital_Good::Capital_Good(Capital_Firm_Agent *  seller_pointer, float seller_price, long long seller_quantity,int life_span)
 {
     pSeller = seller_pointer;
     price = seller_price;
